refactor(Appointment) : 패키지 구조 정리 및 응답 DTO 패키지 변경
- Appointment 관련 응답 DTO를 `recommendation` 패키지로 이동하여 구조 개선
- 기존 `org.noostak.appointment.dto.response`에서 `org.noostak.appointment.dto.response.recommendation`으로 변경
- 관련된 서비스, 컨트롤러, 인터페이스에서 변경된 패키지를 반영하여 코드 수정
