+++ src/main/java/org/noostak/appointment/application/recommendation/AppointmentOptionCommandService.java
@@ -1,7 +1,7 @@
 package org.noostak.appointment.application.recommendation;
 
 import org.noostak.appointment.domain.Appointment;
-import org.noostak.appointment.dto.response.AppointmentOptionAvailabilityResponse;
+import org.noostak.appointment.dto.response.recommendation.AppointmentOptionAvailabilityResponse;
 import org.noostak.appointmentoption.domain.AppointmentOption;
 
 import java.util.List;
