+++ src/main/java/org/noostak/appointment/application/recommendation/AppointmentRecommendationFacade.java
@@ -3,8 +3,8 @@ package org.noostak.appointment.application.recommendation;
 import lombok.RequiredArgsConstructor;
 import org.noostak.appointment.domain.Appointment;
 import org.noostak.appointment.util.TimeSlot;
-import org.noostak.appointment.dto.response.AppointmentOptionAvailabilityResponse;
-import org.noostak.appointment.dto.response.AppointmentPriorityGroupResponse;
+import org.noostak.appointment.dto.response.recommendation.AppointmentOptionAvailabilityResponse;
+import org.noostak.appointment.dto.response.recommendation.AppointmentPriorityGroupResponse;
 import org.noostak.appointment.util.AppointmentAvailabilityMatcher;
 import org.noostak.appointment.util.AppointmentOptionSorter;
 import org.noostak.appointmentmember.domain.AppointmentMemberAvailableTime;
