+++ src/main/java/org/noostak/appointment/application/recommendation/AppointmentRecommendedOptionService.java
@@ -1,6 +1,6 @@
 package org.noostak.appointment.application.recommendation;
 
-import org.noostak.appointment.dto.response.AppointmentRecommendedOptionsResponse;
+import org.noostak.appointment.dto.response.recommendation.AppointmentRecommendedOptionsResponse;
 
 public interface AppointmentRecommendedOptionService {
     AppointmentRecommendedOptionsResponse getRecommendedAppointmentOptions(Long memberId, Long appointmentId);
